change saving code of parameter and change datatype double to bigdecimal in workingparameter

+++ api-loans/src/main/java/com/capitaworld/service/loans/model/corporate/WorkingCapitalParameterRequest.java
@@ -2,6 +2,7 @@ package com.capitaworld.service.loans.model.corporate;
 
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import com.capitaworld.service.loans.model.ProductMasterRequest;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -65,29 +66,29 @@ public class WorkingCapitalParameterRequest extends ProductMasterRequest impleme
 
 	private Integer maxAgeEstablishment;
 
-	private Double maxCollateral;
+	private BigDecimal maxCollateral;
 
-	private Double maxDebtEquity;
+	private BigDecimal maxDebtEquity;
 
-	private Double maxInvestSize;
+	private BigDecimal maxInvestSize;
 
-	private Double maxNetworth;
+	private BigDecimal maxNetworth;
 
-	private Double maxPastTurnover;
+	private BigDecimal maxPastTurnover;
 
 	private Integer maxTenure;
 
 	private Integer minAgeEstablishment;
 
-	private Double minCollateral;
+	private BigDecimal minCollateral;
 
-	private Double minDebtEquity;
+	private BigDecimal minDebtEquity;
 
-	private Double minInvestSize;
+	private BigDecimal minInvestSize;
 
-	private Double minNetworth;
+	private BigDecimal minNetworth;
 
-	private Double minPastTurnover;
+	private BigDecimal minPastTurnover;
 
 	private Integer minTenure;
 
@@ -320,43 +321,43 @@ public class WorkingCapitalParameterRequest extends ProductMasterRequest impleme
 		this.maxAgeEstablishment = maxAgeEstablishment;
 	}
 
-	public Double getMaxCollateral() {
+	public BigDecimal getMaxCollateral() {
 		return this.maxCollateral;
 	}
 
-	public void setMaxCollateral(Double maxCollateral) {
+	public void setMaxCollateral(BigDecimal maxCollateral) {
 		this.maxCollateral = maxCollateral;
 	}
 
-	public Double getMaxDebtEquity() {
+	public BigDecimal getMaxDebtEquity() {
 		return this.maxDebtEquity;
 	}
 
-	public void setMaxDebtEquity(Double maxDebtEquity) {
+	public void setMaxDebtEquity(BigDecimal maxDebtEquity) {
 		this.maxDebtEquity = maxDebtEquity;
 	}
 
-	public Double getMaxInvestSize() {
+	public BigDecimal getMaxInvestSize() {
 		return this.maxInvestSize;
 	}
 
-	public void setMaxInvestSize(Double maxInvestSize) {
+	public void setMaxInvestSize(BigDecimal maxInvestSize) {
 		this.maxInvestSize = maxInvestSize;
 	}
 
-	public Double getMaxNetworth() {
+	public BigDecimal getMaxNetworth() {
 		return this.maxNetworth;
 	}
 
-	public void setMaxNetworth(Double maxNetworth) {
+	public void setMaxNetworth(BigDecimal maxNetworth) {
 		this.maxNetworth = maxNetworth;
 	}
 
-	public Double getMaxPastTurnover() {
+	public BigDecimal getMaxPastTurnover() {
 		return this.maxPastTurnover;
 	}
 
-	public void setMaxPastTurnover(Double maxPastTurnover) {
+	public void setMaxPastTurnover(BigDecimal maxPastTurnover) {
 		this.maxPastTurnover = maxPastTurnover;
 	}
 
@@ -376,43 +377,43 @@ public class WorkingCapitalParameterRequest extends ProductMasterRequest impleme
 		this.minAgeEstablishment = minAgeEstablishment;
 	}
 
-	public Double getMinCollateral() {
+	public BigDecimal getMinCollateral() {
 		return this.minCollateral;
 	}
 
-	public void setMinCollateral(Double minCollateral) {
+	public void setMinCollateral(BigDecimal minCollateral) {
 		this.minCollateral = minCollateral;
 	}
 
-	public Double getMinDebtEquity() {
+	public BigDecimal getMinDebtEquity() {
 		return this.minDebtEquity;
 	}
 
-	public void setMinDebtEquity(Double minDebtEquity) {
+	public void setMinDebtEquity(BigDecimal minDebtEquity) {
 		this.minDebtEquity = minDebtEquity;
 	}
 
-	public Double getMinInvestSize() {
+	public BigDecimal getMinInvestSize() {
 		return this.minInvestSize;
 	}
 
-	public void setMinInvestSize(Double minInvestSize) {
+	public void setMinInvestSize(BigDecimal minInvestSize) {
 		this.minInvestSize = minInvestSize;
 	}
 
-	public Double getMinNetworth() {
+	public BigDecimal getMinNetworth() {
 		return this.minNetworth;
 	}
 
-	public void setMinNetworth(Double minNetworth) {
+	public void setMinNetworth(BigDecimal minNetworth) {
 		this.minNetworth = minNetworth;
 	}
 
-	public Double getMinPastTurnover() {
+	public BigDecimal getMinPastTurnover() {
 		return this.minPastTurnover;
 	}
 
-	public void setMinPastTurnover(Double minPastTurnover) {
+	public void setMinPastTurnover(BigDecimal minPastTurnover) {
 		this.minPastTurnover = minPastTurnover;
 	}
